A fix for InflateException thrown in FrameLayout.
This addresses a bug when ApplicationContext.setThememethod may ignore (system) default theme. The bugexposes itself as an InflateException being thrownon attempt to launch either Dialer or Contactsapp. That happens because (our version of) the apps
use styles with no parents, which do not define
windowContentOverlay attribute value.
Part of the fix is to remove redundant setTheme call.CR: JoshG.