Redo a fix for ESPRESSO-9.
Turns out my fix for ESPRESSO-8 was not quite right,
and has to be rolled back. As a result, the code which
ESPRESSO-8 fix has added to AlertController must be
moved to GlobalActions class. Which makes sense, since
Power Off is the only dialog launched from system context
(as far as we know).

CR:JoshG.