A fix for ESPRESSO-174.
I missed this scenario when was fixing ESPRESSO-9 bug.
This check in makes sure that we pass theme (aware) context
to the confirmation dialog as well.

The dialog is actually invoked in one more place.
Fortunately, this place (in PowerDialog class) can be ignored,
since the parent class is marked as deprecated with
suggestion to use GlobalActions class instead.

CR:JoshG.